Refactor cluster_name change and remove dependency on clustername.GetClusterName from kubelet package.
As part of this commit moved the HostnameProvider to hostinfo package to avoid import cycle.
