+++ pkg/metadata/host/host_tags.go
@@ -8,7 +8,6 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/util/docker"
 	"github.com/DataDog/datadog-agent/pkg/util/ec2"
 	"github.com/DataDog/datadog-agent/pkg/util/gce"
-	"github.com/DataDog/datadog-agent/pkg/util/kubernetes/clustername"
 	k8s "github.com/DataDog/datadog-agent/pkg/util/kubernetes/hostinfo"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
@@ -62,9 +61,9 @@ func getHostTags() *tags {
 		}
 	}
 
-	clusterName := clustername.GetClusterName()
+	clusterName := k8s.GetClusterName()
 	if len(clusterName) != 0 {
-		hostTags = appendToHostTags(hostTags, []string{"cluster-name:" + clusterName})
+		hostTags = appendToHostTags(hostTags, []string{"cluster_name:" + clusterName})
 	}
 
 	k8sTags, err := k8s.GetTags()
