+++ pkg/util/hostname.go
@@ -10,9 +10,7 @@ import (
 	"fmt"
 	"net"
 	"os"
-	"regexp"
 	"runtime"
-	"strings"
 
 	"github.com/DataDog/datadog-agent/pkg/metadata/inventories"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
@@ -22,18 +20,10 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/util/ec2"
 	"github.com/DataDog/datadog-agent/pkg/util/ecs"
 	"github.com/DataDog/datadog-agent/pkg/util/hostname"
+	"github.com/DataDog/datadog-agent/pkg/util/hostname/validate"
 )
 
-const maxLength = 255
-
 var (
-	validHostnameRfc1123 = regexp.MustCompile(`^(([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9\-]*[a-zA-Z0-9])\.)*([A-Za-z0-9]|[A-Za-z0-9][A-Za-z0-9\-]*[A-Za-z0-9])$`)
-	localhostIdentifiers = []string{
-		"localhost",
-		"localhost.localdomain",
-		"localhost6.localdomain6",
-		"ip6-localhost",
-	}
 	hostnameExpvars  = expvar.NewMap("hostname")
 	hostnameProvider = expvar.String{}
 	hostnameErrors   = expvar.Map{}
@@ -43,34 +33,6 @@ func init() {
 	hostnameErrors.Init()
 	hostnameExpvars.Set("provider", &hostnameProvider)
 	hostnameExpvars.Set("errors", &hostnameErrors)
-}
-
-// ValidHostname determines whether the passed string is a valid hostname.
-// In case it's not, the returned error contains the details of the failure.
-func ValidHostname(hostname string) error {
-	if hostname == "" {
-		return fmt.Errorf("hostname is empty")
-	} else if isLocal(hostname) {
-		return fmt.Errorf("%s is a local hostname", hostname)
-	} else if len(hostname) > maxLength {
-		log.Errorf("ValidHostname: name exceeded the maximum length of %d characters", maxLength)
-		return fmt.Errorf("name exceeded the maximum length of %d characters", maxLength)
-	} else if !validHostnameRfc1123.MatchString(hostname) {
-		log.Errorf("ValidHostname: %s is not RFC1123 compliant", hostname)
-		return fmt.Errorf("%s is not RFC1123 compliant", hostname)
-	}
-	return nil
-}
-
-// check whether the name is in the list of local hostnames
-func isLocal(name string) bool {
-	name = strings.ToLower(name)
-	for _, val := range localhostIdentifiers {
-		if val == name {
-			return true
-		}
-	}
-	return false
 }
 
 // Fqdn returns the FQDN for the host if any
@@ -146,7 +108,7 @@ func GetHostnameData() (HostnameData, error) {
 
 	// try the name provided in the configuration file
 	configName := config.Datadog.GetString("hostname")
-	err = ValidHostname(configName)
+	err = validate.ValidHostname(configName)
 	if err == nil {
 		hostnameData := saveHostnameData(cacheHostnameKey, configName, HostnameProviderConfiguration)
 		if !isHostnameCanonicalForIntake(configName) && !config.Datadog.GetBool("hostname_force_config_as_canonical") {
@@ -233,7 +195,7 @@ func GetHostnameData() (HostnameData, error) {
 		if ecs.IsECSInstance() || ec2.IsDefaultHostname(hostName) {
 			instanceID, err := getEC2Hostname()
 			if err == nil {
-				err = ValidHostname(instanceID)
+				err = validate.ValidHostname(instanceID)
 				if err == nil {
 					hostName = instanceID
 					provider = "aws"
