+++ pkg/util/hostname/validate/validate_test.go
@@ -2,8 +2,7 @@
 // under the Apache License Version 2.0.
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-2019 Datadog, Inc.
-
-package util
+package validate
 
 import (
 	"strings"
