+++ pkg/util/hostname_docker.go
@@ -12,6 +12,7 @@ package util
 import (
 	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/util/hostname"
+	"github.com/DataDog/datadog-agent/pkg/util/hostname/validate"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
 
@@ -22,7 +23,7 @@ func getContainerHostname() (bool, string) {
 	if getKubeHostname, found := hostname.ProviderCatalog["kube_apiserver"]; found {
 		log.Debug("GetHostname trying Kubernetes trough API server...")
 		name, err := getKubeHostname()
-		if err == nil && ValidHostname(name) == nil {
+		if err == nil && validate.ValidHostname(name) == nil {
 			return true, name
 		}
 	}
@@ -37,7 +38,7 @@ func getContainerHostname() (bool, string) {
 	log.Debug("GetHostname trying Docker API...")
 	if getDockerHostname, found := hostname.ProviderCatalog["docker"]; found {
 		name, err := getDockerHostname()
-		if err == nil && ValidHostname(name) == nil {
+		if err == nil && validate.ValidHostname(name) == nil {
 			return true, name
 		}
 	}
@@ -49,7 +50,7 @@ func getContainerHostname() (bool, string) {
 	if getKubeletHostname, found := hostname.ProviderCatalog["kubelet"]; found {
 		log.Debug("GetHostname trying Kubernetes trough kubelet API...")
 		name, err := getKubeletHostname()
-		if err == nil && ValidHostname(name) == nil {
+		if err == nil && validate.ValidHostname(name) == nil {
 			return true, name
 		}
 	}
